add scripts for build processes

+++ scripts/build-container.sh
@@ -1,6 +1,15 @@
 SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
-SCRIPT_NAME='build-app.sh'
+SCRIPT_NAME='build-container.sh'
 source "$SCRIPT_DIR/util/logging.sh"
+
+if [ ! -f "$SCRIPT_DIR/../env/container.env" ]
+then
+    log 'No Environment File Detected' $SCRIPT_NAME
+    log 'Initializing New Environment File' $SCRIPT_NAME
+    cp $SCRIPT_DIR/../env/.sample.env $SCRIPT_DIR/../env/container.env
+    log "Please Configure \e[3m$SCRIPT_DIR/../env/container.env\e[0m And Then Reinvoke This Script" $SCRIPT_NAME
+    exit 0
+fi
 
 log 'Removing Running Containers' $SCRIPT_NAME
 docker-compose down
